refactor(transfer): decompose reorder_buffer.rs (SPL-28 #2456)
Splits the 1148-line reorder_buffer.rs into focused submodules under
crates/transfer/src/reorder_buffer/. Public API is unchanged.

- mod.rs: module doc, struct, stats, error, Debug impl
- state.rs: constructors and read-only accessors
- insert.rs: window-admission insert path
- drain.rs: consecutive-drain helper
- tests/{mod,property,metrics,passthrough}.rs: split test surfaces
